Improve graphics services

+++ src/font.rs
@@ -11,10 +11,10 @@ pub struct FontDriver<'a> {
 
 impl<'a> FontDriver<'_> {
     pub fn glyph_for(&self, c: char) -> Option<&[u8]> {
-        let c = c as u32;
+        let c = c as usize;
         if c > 0x20 && c < 0x80 {
             let delta = (self.size.width as usize + 7) / 8 * self.size.height as usize;
-            let base = 0x11 + delta * c as usize;
+            let base = 0x11 + delta * c;
             Some(&self.data[base..base + delta])
         } else {
             None
